Inline DCO decoder test fixtures (mechanical)
Method: apply_patch moved test_utils.rs verbatim into the existing cfg(test) module in ty/mod.rs.
